refactor: Separate post related queries/hooks/component
