+++ src/features/posts/components/PostModal.tsx
@@ -4,8 +4,8 @@ import { useMutation, useQueryClient } from "@tanstack/react-query";
 import { deleteLike, postLike } from "../../core/api/queries";
 import useLikeQuery from "../../core/hooks/useLikeQuery";
 import usePostQuery from "../hooks/usePostQuery";
-import { postDetailsLoader } from "../../core/api/loaders";
-import Post from "../../core/types/post";
+import { postDetailsLoader } from "../../explore/api/loaders";
+import Post from "../../posts/types/post";
 import Modal from "../../core/components/Modal";
 import UserBar from "./UserBar";
 import Timestamp from "../../core/components/Timestamp";
